requirements.txt
pandas==1.5.3
numpy==1.23.5
psycopg2-binary==2.9.5
google-auth==2.17.3
google-api-python-client==2.80.0
openpyxl==3.1.2
openpyxl-image-loader==1.0.5
python-dotenv==1.0.0